# api-doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Summarizer API",
    "version": "0.1.0"
  },
  "paths": {
    "/documents/": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Create Document",
        "operationId": "create_document_documents__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_documents__get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentRead"
                  },
                  "title": "Response List Documents Documents  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_uuid}/": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "operationId": "get_document_documents__document_uuid___get",
        "parameters": [
          {
            "name": "document_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/health": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Redis Health",
        "operationId": "redis_health_documents_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Redis Health Documents Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DocumentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "name",
          "url"
        ],
        "title": "DocumentCreate"
      },
      "DocumentRead": {
        "properties": {
          "document_uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Document Uuid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "data_progress": {
            "type": "number",
            "title": "Data Progress",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "document_uuid",
          "name",
          "url",
          "status",
          "data_progress"
        ],
        "title": "DocumentRead"
      },
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PROCESSING",
          "SUCCESS",
          "FAILED"
        ],
        "title": "DocumentStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}